+++ src/NzbDrone.Core/MetadataSource/MetadataRequestBuilder.cs
@@ -1,6 +1,8 @@
 using NzbDrone.Common.Extensions;
 using NzbDrone.Common.Http;
 using NzbDrone.Core.Configuration;
+using NzbDrone.Core.Languages;
+using NzbDrone.Core.Parser;
 
 namespace NzbDrone.Core.MetadataSource
 {
@@ -20,7 +22,31 @@ namespace NzbDrone.Core.MetadataSource
 
         public IHttpRequestBuilderFactory GetRequestBuilder()
         {
-            return new HttpRequestBuilder(_configService.MetadataSource.TrimEnd("/") + "/{route}").KeepAlive().CreateFactory();
+            var builder = new HttpRequestBuilder(_configService.MetadataSource.TrimEnd("/") + "/{route}");
+
+            var languageHeader = GetUiLanguageTag();
+            if (languageHeader.IsNotNullOrWhiteSpace())
+            {
+                builder.SetHeader("Accept-Language", languageHeader);
+            }
+
+            return builder.KeepAlive().CreateFactory();
+        }
+
+        private string GetUiLanguageTag()
+        {
+            var isoLanguage = IsoLanguages.Get((Language)_configService.UILanguage) ?? IsoLanguages.Get(Language.English);
+            if (isoLanguage == null || isoLanguage.TwoLetterCode.IsNullOrWhiteSpace())
+            {
+                return null;
+            }
+
+            if (isoLanguage.CountryCode.IsNullOrWhiteSpace())
+            {
+                return isoLanguage.TwoLetterCode;
+            }
+
+            return $"{isoLanguage.TwoLetterCode}-{isoLanguage.CountryCode.ToUpperInvariant()}";
         }
     }
 }